feat: Added notification to exit chat mode when executing another command
This commit adds a new feature to the chat mode where, if the user executes another command while in chat mode, they will be notified to exit chat mode before proceeding with the requested action. This helps to prevent errores and confusion when executing commands that are not valid in chat mode.